backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.36.0
dotenv==0.9.9
flask==3.1.1
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-jwt-extended==4.7.1
flask-mail==0.9.1
flask-sqlalchemy==3.1.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.31.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyjwt==2.10.1
pymysql==1.1.1
pyparsing==3.2.3
python-dotenv==1.0.0
pytz==2025.2
redis==4.5.5
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlalchemy==2.0.41
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
valkey==6.0.2
werkzeug==3.1.3
gunicorn==21.2.0
